srcs/requirements.txt
ftfy 
regex 
tqdm
torch
torchvision
git+https://github.com/openai/CLIP.git
Pillow
numpy
opencv-python
scikit-image
scikit-learn
